refactor(event): refactor by code review
