+++ packages/event/src/error/index.ts
@@ -1,3 +1,3 @@
-export * from './pub-sub.error'
-export * from './validation-error'
+export * from './publish-pub-sub.error'
+export * from './validation-attribute.error'
 export * from './client-not-found.error'
