+++ packages/event/src/index.ts
@@ -1,4 +1,3 @@
-export * from './service/send-event.service'
-export * from './enum/pub-sub.enum'
-export * from './enum/client-event.enum'
-export * from './dto/pub-sub-attribute.dto'
+export * from './service'
+export * from './enum'
+export * from './dto'
